Allow linking to specific lines.
This is slightly hacky, but seems like a useful thing to be able to do
for any online version of a play - be able to link to "Tomorrow, and
tomorrow, and tomorrow, creeps in this petty pace" and so on.

I've replaced the `<br/>` tags with spans, so that the line links can
be shown only when hovering over a line. I don't think this should
break anything and it still looks good in my browser at least.